color wheel ui improvements
