--- express/code/blocks/color-wheel-alt/color-wheel-alt.css
@@ -1,189 +1,0 @@
-@import url('../../scripts/color-shared/components/strips/color-strip.css');
-
-.color-wheel.ax-shell-host {
-  width: 100%;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-.color-wheel-alt-strip-host {
-  height: 400px;
-  min-height: 400px;
-  overflow: auto;
-}
-
-/* Sidebar + tabs: allow column to shrink so the wheel measures against real width */
-.color-wheel-alt .ax-shell-slot--sidebar {
-  min-width: 0;
-}
-
-.color-wheel-alt .color-wheel-content {
-  min-width: 0;
-  max-width: 100%;
-  box-sizing: border-box;
-}
-
-/* Harmony picker — WCAG-oriented layout */
-.color-wheel-harmony-section {
-  margin-bottom: 1rem;
-  min-width: 0;
-}
-
-.color-wheel-harmony-title-row {
-  display: flex;
-  flex-wrap: wrap;
-  align-items: baseline;
-  gap: 0.35rem 0.75rem;
-  margin-bottom: 0.75rem;
-}
-
-.color-wheel-harmony-title-static {
-  font-weight: 600;
-  font-size: 1rem;
-  color: var(--spectrum-gray-800, #2c2c2c);
-}
-
-.color-wheel-harmony-current-name {
-  font-weight: 500;
-  font-size: 1rem;
-  color: var(--spectrum-gray-700, #464646);
-}
-
-.color-wheel-sr-only {
-  position: absolute;
-  width: 1px;
-  height: 1px;
-  padding: 0;
-  margin: -1px;
-  overflow: hidden;
-  clip: rect(0, 0, 0, 0);
-  white-space: nowrap;
-  border: 0;
-}
-
-.color-wheel-harmony-toolbar {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  gap: 0.25rem;
-  width: 100%;
-  min-width: 0;
-}
-
-.color-wheel-harmony-scroll-region {
-  flex: 1 1 0;
-  min-width: 0;
-  overflow-x: auto;
-  overflow-y: hidden;
-  scroll-snap-type: x proximity;
-  scrollbar-width: thin;
-  scrollbar-color: var(--spectrum-gray-400, #999) transparent;
-}
-
-@media (prefers-reduced-motion: reduce) {
-  .color-wheel-harmony-scroll-region {
-    scroll-behavior: auto;
-  }
-}
-
-.color-wheel-harmony-radiogroup {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: nowrap;
-  gap: 4px;
-  width: max-content;
-  padding: 2px 0;
-  align-items: center;
-}
-
-.color-wheel-harmony-option {
-  box-sizing: border-box;
-  width: 48px;
-  height: 48px;
-  min-width: 48px;
-  min-height: 48px;
-  padding: 0;
-  margin: 0;
-  border: 2px solid var(--spectrum-gray-300, #d3d3d3);
-  border-radius: 6px;
-  background: var(--spectrum-gray-75, #fafafa);
-  cursor: pointer;
-  flex-shrink: 0;
-  scroll-snap-align: start;
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  transition: border-color 0.15s ease, box-shadow 0.15s ease;
-}
-
-.color-wheel-harmony-option:hover {
-  border-color: var(--spectrum-gray-500, #707070);
-  background: var(--spectrum-gray-100, #f5f5f5);
-}
-
-.color-wheel-harmony-option[aria-checked='true'] {
-  border-color: var(--spectrum-blue-900, #0265dc);
-  box-shadow: 0 0 0 1px var(--spectrum-blue-900, #0265dc);
-  background: var(--spectrum-blue-100, #e0f2ff);
-}
-
-.color-wheel-harmony-option:focus {
-  outline: none;
-}
-
-.color-wheel-harmony-option:focus-visible {
-  outline: 3px solid var(--spectrum-blue-900, #0265dc);
-  outline-offset: 2px;
-  z-index: 1;
-}
-
-.color-wheel-harmony-option img {
-  display: block;
-  width: 48px;
-  height: 48px;
-  border-radius: 4px;
-  pointer-events: none;
-}
-
-.color-wheel-harmony-scroll {
-  box-sizing: border-box;
-  flex: 0 0 auto;
-  width: 48px;
-  height: 48px;
-  min-width: 48px;
-  min-height: 48px;
-  padding: 0;
-  margin: 0;
-  border: 2px solid var(--spectrum-gray-400, #909090);
-  border-radius: 6px;
-  background: var(--spectrum-gray-75, #fafafa);
-  color: var(--spectrum-gray-800, #2c2c2c);
-  cursor: pointer;
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  font-size: 1.5rem;
-  line-height: 1;
-}
-
-.color-wheel-harmony-scroll:hover:not(:disabled) {
-  background: var(--spectrum-gray-200, #e8e8e8);
-  border-color: var(--spectrum-gray-600, #5c5c5c);
-}
-
-.color-wheel-harmony-scroll:focus {
-  outline: none;
-}
-
-.color-wheel-harmony-scroll:focus-visible {
-  outline: 3px solid var(--spectrum-blue-900, #0265dc);
-  outline-offset: 2px;
-}
-
-.color-wheel-harmony-scroll[hidden] {
-  display: none !important;
-}
-
-.mobile-container {
-  max-width: 343px;
-}
