+++ express/code/blocks/color-wheel/color-wheel.css
@@ -1,16 +1,24 @@
 @import url('../../scripts/color-shared/components/strips/color-strip.css');
 
-.color-wheel {
+.color-wheel.ax-shell-host {
   width: 100%;
-  padding: var(--spacing-m);
   box-sizing: border-box;
 }
 
-.palette-builder {
-  padding: var(--spacing-m);
+.color-wheel .ax-shell-slot--sidebar {
+  min-width: 0;
 }
 
-/* Harmony picker — WCAG-oriented layout */
+.color-wheel .color-wheel-content {
+  min-width: 0;
+  max-width: 100%;
+  box-sizing: border-box;
+}
+
+.color-wheel color-wheel-express {
+  padding-block: var(--spacing-300) var(--spacing-100);
+}
+
 .color-wheel-harmony-section {
   margin-bottom: 1rem;
   min-width: 0;
@@ -20,7 +28,7 @@
   display: flex;
   flex-wrap: wrap;
   align-items: baseline;
-  gap: 0.35rem 0.75rem;
+  gap: 2px;
   margin-bottom: 0.75rem;
 }
 
@@ -67,12 +75,6 @@
   scrollbar-color: var(--spectrum-gray-400, #999) transparent;
 }
 
-@media (prefers-reduced-motion: reduce) {
-  .color-wheel-harmony-scroll-region {
-    scroll-behavior: auto;
-  }
-}
-
 .color-wheel-harmony-radiogroup {
   display: flex;
   flex-direction: row;
@@ -87,35 +89,26 @@
   box-sizing: border-box;
   width: 48px;
   height: 48px;
-  min-width: 48px;
-  min-height: 48px;
   padding: 0;
   margin: 0;
-  border: 2px solid var(--spectrum-gray-300, #d3d3d3);
-  border-radius: 6px;
-  background: var(--spectrum-gray-75, #fafafa);
+  border-radius: var(--Radius-corner-radius-75);
+  border: 1px solid transparent;
   cursor: pointer;
   flex-shrink: 0;
   scroll-snap-align: start;
   display: inline-flex;
   align-items: center;
   justify-content: center;
-  transition: border-color 0.15s ease, box-shadow 0.15s ease;
+  transition: var(--standard-transition-hover-out-border-color);
 }
 
-.color-wheel-harmony-option:hover {
-  border-color: var(--spectrum-gray-500, #707070);
-  background: var(--spectrum-gray-100, #f5f5f5);
+.color-wheel-harmony-option:hover:not([aria-checked='true']) {
+  border-color: var(--color-black);
+  transition: var(--standard-transition-hover-in-border-color);
 }
 
 .color-wheel-harmony-option[aria-checked='true'] {
-  border-color: var(--spectrum-blue-900, #0265dc);
-  box-shadow: 0 0 0 1px var(--spectrum-blue-900, #0265dc);
-  background: var(--spectrum-blue-100, #e0f2ff);
-}
-
-.color-wheel-harmony-option:focus {
-  outline: none;
+  border-color: var(--Background-accent-default);
 }
 
 .color-wheel-harmony-option:focus-visible {
@@ -126,8 +119,8 @@
 
 .color-wheel-harmony-option img {
   display: block;
-  width: 48px;
-  height: 48px;
+  width: 46px;
+  height: 46px;
   border-radius: 4px;
   pointer-events: none;
 }
@@ -173,4 +166,28 @@
 
 .mobile-container {
   max-width: 343px;
+}
+
+@media (prefers-reduced-motion: reduce) {
+  .color-wheel-harmony-scroll-region {
+    scroll-behavior: auto;
+  }
+}
+
+@media (min-width: 1200px) {
+  .color-wheel {
+    padding: var(--spacing-600);
+    max-width: var(--block-wd-max-width);
+    margin: auto;
+  }
+
+  .color-wheel-strip-host {
+    height: 600px;
+    min-height: 600px;
+    overflow: auto;
+  }
+
+  sp-theme {
+    height: 100%;
+  }
 }
