+++ express/code/scripts/widgets/spectrum/src/icons-workflow.js
@@ -12,3 +12,6 @@ import '@spectrum-web-components/icons-workflow/icons/sp-icon-cclibrary.js';
 import '@spectrum-web-components/icons-workflow/icons/sp-icon-chevron-down.js';
 import '@spectrum-web-components/icons-workflow/icons/sp-icon-chevron-left.js';
 import '@spectrum-web-components/icons-workflow/icons/sp-icon-checkmark-circle.js';
+import '@spectrum-web-components/icons-workflow/icons/sp-icon-sampler.js';
+import '@spectrum-web-components/icons-workflow/icons/sp-icon-image.js';
+import '@spectrum-web-components/icons-workflow/icons/sp-icon-color-wheel.js';
